+++ addon/initializers/local-storage-adapter.js
@@ -1,17 +1,15 @@
 import DS from 'ember-data';
-import Adapter from 'ember-local-storage/adapters/adapter';
+import { importData, exportData } from 'ember-local-storage/helpers/import-export';
 
 export function initialize() {
   if (!DS.Store.prototype._emberLocalStoragePatched) {
-    const adapter = Adapter.create();
-
     DS.Store.reopen({
       _emberLocalStoragePatched: true,
       importData: function(json, options) {
-        return adapter.importData.call(adapter, this, json, options);
+        return importData(this, json, options);
       },
       exportData: function(types, options) {
-        return adapter.exportData.call(adapter, this, types, options);
+        return exportData(this, types, options);
       }
     });
   }
