+++ tests/unit/adapters/indices-test.js
@@ -1,25 +1,15 @@
-import Ember from 'ember';
 import { moduleFor, test } from 'ember-qunit';
 import {
   storageEqual,
   storageDeepEqual
 } from '../../helpers/storage';
 
-const {
-  getOwner
-} = Ember;
-
 moduleFor('adapter:application', 'Unit | Adapter | indices', {
   // Specify the other units that are required for this test.
   // needs: ['serializer:foo']
   beforeEach: function() {
-    const adapter = getOwner(this).lookup('adapter:application');
-
-    ['projects'].forEach(function(key) {
-      adapter._getIndex(key).reset();
-    });
-
     window.localStorage.clear();
+    window.sessionStorage.clear();
   }
 });
