+++ tests/unit/models/query-record-test.js
@@ -3,7 +3,6 @@ import { moduleForModel, test } from 'ember-qunit';
 
 const {
   get,
-  getOwner,
   run
 } = Ember;
 
@@ -23,13 +22,8 @@ moduleForModel('post', 'Unit | Model | queryRecord', {
     'model:task'
   ],
   beforeEach: function() {
-    const adapter = getOwner(this).lookup('adapter:application');
-
-    ['posts', 'users', 'projects', 'comments', 'pets'].forEach(function(key) {
-      adapter._getIndex(key).reset();
-    });
-
     window.localStorage.clear();
+    window.sessionStorage.clear();
   }
 });
